refactor(settings/copilot): 提取 StatusCard 为独立组件
